+++ server/__tests__/server.test.js
@@ -1,0 +1,13 @@
+const mongoose = require('mongoose')
+const request = require('supertest')
+const faker = require('faker')
+const server = require('../server')
+
+const date = faker.date
+const { words, word } = faker.random
+
+describe('server', () => {
+  it('runs a test', () => {
+    expect(server).toBeTruthy()
+  })
+})
